[banshee] disable B-C-E dependencies for now
The compilation of fftw dependency is failing in
Mavericks[0], and libsdl would depend on XQuartz,
which is still pending an investigation to check
if it can be built inside bockbuild itself or we
need to state it as a proper binary dependency
for creating the Mac bundles.

So, let's postpone this for now, and disable the
dependencies for B-C-E for the moment, as B-C-E
is lower priority than Banshee Core.

[0] Undefined symbols for architecture i386:
  "_main", referenced from:
      start in crt1.10.6.o
ld: symbol(s) not found for architecture i386
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [bench] Error 1
make[1]: *** [all-recursive] Error 1
make: *** [all] Error 2