[libogg] Fix libogg build by upgrading it to 1.3.2
Build for libogg was failing with this warnings and
error:

aclocal: warning: autoconf input should be named 'configure.ac', not 'configure.in'
configure.in:18: warning: macro 'AM_PROG_LIBTOOL' not found in library
configure.in:22: warning: macro 'AM_CONFIG_HEADER' not found in library
configure.in:18: error: possibly undefined macro: AM_PROG_LIBTOOL
      If this token and others are legitimate, please use m4_pattern_allow.
      See the Autoconf documentation.
configure.in:22: error: possibly undefined macro: AM_CONFIG_HEADER

By upgrading to 1.3.2 we fix this and we don't need a custom
patch anymore.

The only profile that was using libogg was banshee's.