[solitary] Ensure main assemblies are loaded in ctor (bgo#751339)
A similar problem to the first encountered in the bug referenced
above: Assembly.Load(rname.FullName) could be throwing a
FileNotFoundException() when starting to call Load() on items,
because, for example, Banshee.Osx.dll depends on MonoMac.dll,
but the latter isn't in the same directory as the former.

The way to fix this is to ensure we load each main assembly first,
when constructing their corresponding AssemblyItem() instance
(in the temp path, as done in previous commit), and then their
dependencies are loaded when their Load() methods start being
called (so the .NET/Mono framework would detect that MonoMac.dll
has already been loaded previously, when Assembly.Load(fullname)
is called, and would reuse its Assembly instance), and there's
no path information in referenced assemblies, just AssemblyNames.

This makes the loading algorithm not be a full DFS strategy,
because the first level of assemblies would be loaded first
(BFS search for the first level, DFS for the dependencies of
each of the assemblies in the first level).

This also allows the code be a bit cleaner, because we can now
turn Item.File's setter from protected to private, and bubble
up its assignment at ctor time, reducing more mutability.

Fixes last part of https://bugzilla.gnome.org/show_bug.cgi?id=751339